add/manage markers (signs) for proof save points
This adds markers '>' (signs in vim speech) to proof save points and
aims to manage their removal cleverly. A marker is added on any line
where the prove-state is saved `hv`. When the proof state is dropped
`hd`, so are all the markers. When the proof is reverted to the previous
save point `hB` or the previous sequence of tacics is revoked `hb`, all
later markers are removed.